Refactor defaultpreemption for out-of-tree plugins
